+++ CHANGES.rst
@@ -3,6 +3,9 @@ Version 2.2.3
 
 Unreleased
 
+-   Autoescaping is now enabled by default for ``.svg`` files. Inside
+    templates this behavior can be changed with the ``autoescape`` tag.
+    :issue:`4831`
 -   Fix the type of ``template_folder`` to accept ``pathlib.Path``. :issue:`4892`